+++ CHANGELOG.md
@@ -20,6 +20,8 @@
   ([#7361](https://github.com/mitmproxy/mitmproxy/pull/7361), [#7373](https://github.com/mitmproxy/mitmproxy/pull/7373), @mhils)
 - Fix a bug in windows management in mitmproxy TUI whereby the help window does not appear if "?" is pressed within the overlay
   ([#6500](https://github.com/mitmproxy/mitmproxy/pull/6500), @emanuele-em)
+- Copy request/response data to the clipboard in mitmweb
+  ([#7352](https://github.com/mitmproxy/mitmproxy/pull/7352), @lups2000)
 
 ## 24 November 2024: mitmproxy 11.0.1